sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ax-serving"
version = "0.1.0"
requires-python = ">=3.10"
description = "Python SDK for ax-serving: gRPC + OpenAI-compatible inference client"
dependencies = [
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.25",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[tool.hatch.build.targets.wheel]
packages = ["ax_serving"]
